change dtype on portfolio to float64

+++ enums/portfolio.py
@@ -27,6 +27,6 @@ class PortfolioMatrix:
 
 PORTFOLIO_DTYPE = np.dtype([
     (PortfolioMatrix.INDEX, np.int32),
-    (PortfolioMatrix.VOLUME, np.float32),
-    (PortfolioMatrix.USD_VALUE, np.float32),
+    (PortfolioMatrix.VOLUME, np.float64),
+    (PortfolioMatrix.USD_VALUE, np.float64),
 ])
